Refactor dialogs into a single dialog.
* Refactor attaching dialog so that the process list can be dynamically
discovered.
* Add refresh button to reload process list.
* Refactor connection dialog into the attaching dialog.
* Add connection light to clarify is a valid connection has been made,
useful when no processes can be shown.
