+++ src/WacVsTools.Core/AttachToWacProcess/AttachToWacProcessDialog.xaml
@@ -1,17 +1,39 @@
 ﻿<Window x:Class="WacVsTools.Core.AttachToWacProcess.AttachToWacProcessDialog"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="Attach to WAC Process" Height="400" Width="700" WindowStartupLocation="CenterOwner" ResizeMode="NoResize" Background="FloralWhite" >
+        Title="Attach to WAC Process" Height="500" Width="700" WindowStartupLocation="CenterOwner" ResizeMode="NoResize">
     <Window.Resources>
         <ResourceDictionary Source="Resources.xaml"/>
     </Window.Resources>
     <Grid Margin="10">
         <Grid.RowDefinitions>
+            <RowDefinition Height="30*"/>
+            <RowDefinition Height="35*"/>
             <RowDefinition Height="223*"/>
             <RowDefinition Height="46*"/>
             <RowDefinition Height="46*"/>
         </Grid.RowDefinitions>
-        <DataGrid x:Name="Processes" AutoGenerateColumns="False" ItemsSource="{Binding Processes}"
+        <Grid Grid.Row="0" VerticalAlignment="Center" Grid.ColumnSpan="2" Margin="0,7">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="auto" />
+                <ColumnDefinition Width="*" />
+            </Grid.ColumnDefinitions>
+            <TextBlock Grid.Column="0" Text="Connection Type:" Margin="0,0,20,0" />
+            <ComboBox Grid.Column="1" Name="ConnectionType" SelectionChanged="ConnectionType_SelectionChanged" SelectedIndex="0" />
+        </Grid>
+        <Grid Grid.Row="1" VerticalAlignment="Center" Grid.ColumnSpan="2" Margin="0,5">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="auto" />
+                <ColumnDefinition Width="*" />
+                <ColumnDefinition Width="40" />
+            </Grid.ColumnDefinitions>
+            <TextBlock Text="Connection Target:" Margin="0,0,20,0" />
+            <TextBox Grid.Column="1" MinWidth="250" MinHeight="20" IsEnabled="True" Name="ConnectionTarget" TextChanged="ConnectionTarget_TextChanged" />
+            <Canvas Grid.Column="2" Width="20" Height="20">
+                <Ellipse Name="StatusLight" Canvas.Left="0" Canvas.Top="0" Width="20" Height="20" Fill="Red" />
+            </Canvas>
+        </Grid>
+        <DataGrid Grid.Row="2" x:Name="Processes" AutoGenerateColumns="False" ItemsSource="{Binding Processes}"
                   SelectionMode="Extended" SelectionChanged="Processes_SelectionChanged" MouseDoubleClick="Processes_MouseDoubleClick"
                   CanUserAddRows="False" CanUserDeleteRows="False" CanUserResizeRows="False" RowDetailsVisibilityMode="Collapsed">
             <DataGrid.Columns>
@@ -21,7 +43,7 @@
                 <DataGridTextColumn Header="Command Line" Binding="{Binding CommandLine, Mode=OneWay}" Width="*" />
             </DataGrid.Columns>
         </DataGrid>
-        <Grid Grid.Row="1" VerticalAlignment="Center">
+        <Grid Grid.Row="3" VerticalAlignment="Center">
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="auto" />
                 <ColumnDefinition Width="*" />
@@ -33,13 +55,14 @@
                 <AccessText>_Select...</AccessText>
             </Button>
         </Grid>
-        <Grid Grid.Row="2" VerticalAlignment="Center">
+        <Grid Grid.Row="4" VerticalAlignment="Center">
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="*" />
                 <ColumnDefinition Width="auto" />
             </Grid.ColumnDefinitions>
             <TextBlock x:Name="NoRecordsError" FontWeight="Bold" HorizontalAlignment="Left" Text="It's empty here. Are you sure WAC is running on your box?" Foreground="Red" IsEnabled="False" />
             <StackPanel Grid.Column="1" Orientation="Horizontal" HorizontalAlignment="Right">
+                <Button Name="btnRefresh" Click="btnRefresh_Click" IsDefault="True" IsEnabled="False" VerticalAlignment="Top">Refresh</Button>
                 <Button Name="btnOk" Click="btnOk_Click" IsDefault="True" IsEnabled="False">Attach</Button>
                 <Button Name="btnCancel" Click="btnCancel_Click" IsCancel="True">Cancel</Button>
             </StackPanel>
