+++ src/WacVsTools.Core/AttachToWacProcess/AttachToWacProcessDialog.xaml.cs
@@ -3,13 +3,17 @@
     using System;
     using System.Collections.Generic;
     using System.Linq;
+    using System.Net.NetworkInformation;
+    using System.Threading;
     using System.Windows;
     using System.Windows.Controls;
+    using System.Windows.Media;
 
     public partial class AttachToWacProcessDialog : Window
     {
         private AttachToWacProcessDialogModel model;
         private HashSet<int> selectedProcesses;
+        private TypingAssistant assistant;
 
         public AttachToWacProcessDialog(AttachToWacProcessDialogModel model)
         {
@@ -20,9 +24,12 @@
 
             this.DataContext = this.model;
 
-            InitializeComponent();
+            assistant = new TypingAssistant();
+            assistant.Idled += assistant_Idled;
 
+            InitializeComponent();
             NoRecordsError.Visibility = model.Processes.Count == 0 ? Visibility.Visible : Visibility.Hidden;
+            ConnectionType.ItemsSource = model.ConnectionTypes;
         }
 
         private void Processes_SelectionChanged(object sender, SelectionChangedEventArgs e)
@@ -32,6 +39,13 @@
                 selectedProcesses.Remove(removedItem.Id);
 
             btnOk.IsEnabled = selectedProcesses.Count > 0;
+        }
+
+        private void btnRefresh_Click(object sender, RoutedEventArgs e)
+        {
+            model.Processes.Clear();
+            UpdateConnection(validConnection: false);
+            assistant.TextChanged();
         }
 
         private void btnOk_Click(object sender, RoutedEventArgs e)
@@ -64,6 +78,105 @@
         private void SelectEngines_Click(object sender, RoutedEventArgs e)
         {
             model.DebuggerEngines = model.MenuCommands.ShowSelectDebuggerEngineDialog(model.DebuggerEngines.Clone());
+        }
+
+        private void ConnectionType_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            if ((string)ConnectionType.SelectedItem == "Local Machine")
+            {
+                ConnectionTarget.IsEnabled = false;
+                ConnectionTarget.Text = Environment.MachineName;
+                ConnectionTarget.Background = Background.Clone();
+            }
+            else if ((string)ConnectionType.SelectedItem == "Remote Connection")
+            {
+                ConnectionTarget.IsEnabled = true;
+                ConnectionTarget.Text = String.Empty;
+                ConnectionTarget.Background = Brushes.White;
+            }
+        }
+
+        private void ConnectionTarget_TextChanged(object sender, TextChangedEventArgs e)
+        {
+            model.Processes.Clear();
+            UpdateConnection(validConnection: false);
+            if (string.IsNullOrWhiteSpace(ConnectionTarget.Text))
+            {
+                return;
+            }
+
+            string[] address = ConnectionTarget.Text.Split(':');
+            model.Host = address[0];
+            if (address.Length >= 2)
+            {
+                model.Port = address[1];
+            }
+
+            assistant.TextChanged();
+        }
+
+        private void assistant_Idled(object sender, EventArgs e)
+        {
+            string host = model.Host;
+            bool machineExists = false;
+            try
+            {
+                Ping ping = new Ping();
+                PingReply reply = ping.Send(host, timeout: 500);
+
+                if (reply.Status == IPStatus.Success)
+                {
+                    machineExists = true;
+                }
+            }
+            catch (PingException)
+            {
+                machineExists = false;
+            }
+
+            Processes.Dispatcher.Invoke(() =>
+            {
+                if (machineExists && host == model.Host)
+                {
+                    var processes = model.GetWacProcessesFromHost();
+                    if (processes != null)
+                    {
+                        foreach (var process in processes)
+                        {
+                            model.Processes.Add(process);
+                        }
+                    }
+                    UpdateConnection(validConnection: processes != null);
+                }
+            });
+        }
+
+        private void UpdateConnection(bool validConnection)
+        {
+            StatusLight.Fill = validConnection ? Brushes.Green : Brushes.Red;
+            btnRefresh.IsEnabled = validConnection;
+            NoRecordsError.Visibility = validConnection && model.Processes.Count == 0 ? Visibility.Visible : Visibility.Hidden;
+        }
+    }
+
+    // Based on https://stackoverflow.com/questions/33776387/dont-raise-textchanged-while-continuous-typing
+    public class TypingAssistant
+    {
+        public event EventHandler Idled = delegate { };
+        private readonly int timeoutMilliseconds;
+        Timer timer;
+
+        public TypingAssistant(int timeoutMilliseconds = 600)
+        {
+            this.timeoutMilliseconds = timeoutMilliseconds;
+            this.timer = new Timer(p =>
+            {
+                Idled(this, EventArgs.Empty);
+            });
+        }
+        public void TextChanged()
+        {
+            timer.Change(timeoutMilliseconds, Timeout.Infinite);
         }
     }
 }
