--- src/WacVsTools.Core/AttachToWacProcess/ConnectionTypeDialog.xaml
@@ -1,38 +1,0 @@
-<Window x:Class="WacVsTools.Core.AttachToWacProcess.ConnectionTypeDialog"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="Connection" Height="150" Width="700" WindowStartupLocation="CenterOwner" ResizeMode="NoResize" Background="FloralWhite" >
-    <Window.Resources>
-        <ResourceDictionary Source="Resources.xaml"/>
-    </Window.Resources>
-    <Grid Margin="10">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="30*"/>
-            <ColumnDefinition Width="307*"/>
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="*"/>
-        </Grid.RowDefinitions>
-        <Grid Grid.Row="0" VerticalAlignment="Center" Grid.ColumnSpan="2" Margin="0,7">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="auto" />
-                <ColumnDefinition Width="*" />
-            </Grid.ColumnDefinitions>
-            <TextBlock Grid.Column="0" Text="Connection Type:" Margin="0,0,20,0" />
-            <ComboBox Grid.Column="1" Name="ConnectionType" SelectionChanged="ConnectionType_SelectionChanged" SelectedIndex="0" />
-        </Grid>
-        <Grid Grid.Row="1" VerticalAlignment="Center" Grid.ColumnSpan="2" Margin="0,5">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="auto" />
-                <ColumnDefinition Width="*" />
-                <ColumnDefinition Width="auto" />
-            </Grid.ColumnDefinitions>
-            <TextBlock Text="Connection Target:" Margin="0,0,20,0" />
-            <TextBox Grid.Column="1" MinWidth="250" IsEnabled="True" Name="ConnectionTarget" TextChanged="ConnectionTarget_TextChanged" />
-            <Button Grid.Column="2" x:Name="SelectMachine" Click="SelectMachine_Click" IsEnabled="False">
-                <AccessText>_Select...</AccessText>
-            </Button>
-        </Grid>
-    </Grid>
-</Window>
